+++ src/types/index.ts
@@ -4,3 +4,4 @@ export { default as VibeComponent } from "./VibeComponent";
 export { default as VibeComponentProps } from "./VibeComponentProps";
 export * from "./ArrayLastElement";
 export * from "./SplitString";
+export * from "./ElementContent";
